+++ pybamm/models/submodels/interface/sei/__init__.py
@@ -5,3 +5,4 @@ from .reaction_limited import ReactionLimited
 from .solvent_diffusion_limited import SolventDiffusionLimited
 from .electron_migration_limited import ElectronMigrationLimited
 from .interstitial_diffusion_limited import InterstitialDiffusionLimited
+from .ec_reaction_limited import EcReactionLimited