+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/test_spm.py
@@ -148,6 +148,11 @@ class TestSPMWithSEI(unittest.TestCase):
         model = pybamm.lithium_ion.SPM(options)
         model.check_well_posedness()
 
+    def test_well_posed_ec_reaction_limited(self):
+        options = {"sei": "ec reaction limited", "sei porosity change": True}
+        model = pybamm.lithium_ion.SPM(options)
+        model.check_well_posedness()
+
 
 if __name__ == "__main__":
     print("Add -v for more debug output")